MET-2342 Modify type long for content size to Long

+++ metis-media-service/src/main/java/eu/europeana/metis/mediaprocessing/model/AbstractInMemoryFile.java
@@ -39,8 +39,8 @@ abstract class AbstractInMemoryFile implements ResourceRelatedFile {
   }
 
   @Override
-  public long getContentSize() {
-    return content.length;
+  public Long getContentSize() {
+    return (long) content.length;
   }
 
   @Override
